requirements.txt
streamlit
pandas
openpyxl
plotly
GitPython
python-dotenv
